requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
motor==3.4.0
pymongo==4.6.3
